Move `EndpointHashStrategy` to top level and use it for `HealthCheckedEndpointGroup`
